feat(auto_route_generator): add `AutoRouterModuleBuilder`
