+++ auto_route_generator/lib/src/builders/auto_router_module_builder.dart
@@ -1,0 +1,10 @@
+import 'auto_router_builder_base.dart';
+
+class AutoRouterModuleBuilder extends AutoRouterBuilderBase {
+  AutoRouterModuleBuilder({super.options})
+      : super(
+          generatedExtension: '.module.dart',
+          allowSyntaxErrors: true,
+          annotationName: 'AutoRouterConfig.module',
+        );
+}
